+++ src/main/java/vn/htdttt/btl/projection/ResultSet.java
@@ -1,9 +1,9 @@
 package vn.htdttt.btl.projection;
 
 public interface ResultSet {
-    int getRatThap();
-    int getThap();
-    int getTrungBinh();
-    int getCao();
-    int getRatCao();
+    double getRatThap();
+    double getThap();
+    double getTrungBinh();
+    double getCao();
+    double getRatCao();
 }
